Modefied Service class methods to handle null cases, modification in controller class

+++ src/main/java/com/springBootApplication/CRUDApplication/Service/EmpServiceImpl.java
@@ -5,7 +5,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class EmpServiceImpl implements EmpService {
@@ -19,35 +18,32 @@ public class EmpServiceImpl implements EmpService {
     }
 
     @Override
-    public Optional<Employee> getEmpByID(Long id) {
-        return employeeRepo.findById(id);
+    public Employee getEmpByID(Long id) {
+        return employeeRepo.findById(id).orElse(null);
     }
 
 
     @Override
-    public String updateEmp(Employee employee) {
-        if (employeeRepo.existsById(employee.getEmpId())){
-            Employee oldDetails = employeeRepo.findById(employee.getEmpId()).orElse(employee);
-            oldDetails.setName(employee.getName());
-            oldDetails.setPrimarySkill(employee.getPrimarySkill());
-            oldDetails.setSecondarySkill(employee.getSecondarySkill());
-            employeeRepo.save(oldDetails);
-            return "Details updated for Employee -: "+employee.getName();
-        }
-        employeeRepo.save(employee);
-        return "Details updated for Employee: "+employee.getName();
+    public Employee updateEmp(Employee employee) {
+        if(!employeeRepo.existsById(employee.getEmpId())){ return null;}
+        return employeeRepo.save(employee);
     }
 
     @Override
     public String removeEmp(Long id) {
-        employeeRepo.deleteById(id);
-        return "Employee with EmpID: "+id+" is removed to the Database";
+        if(employeeRepo.existsById(id)){
+            employeeRepo.deleteById(id);
+        return "success";}
+        return "failed";
     }
 
     @Override
-    public String addEmp(Employee employee) {
-        employeeRepo.save(employee);
-        return "Employee: "+employee.getName()+" is added to the Database";
+    public Employee addEmp(Employee employee) {
+        if (employeeRepo.existsById(employee.getEmpId())){
+            return null;
+        }
+       return employeeRepo.save(employee);
+
     }
 
 }
